Moved endpoints to work with device to separated controller class
