+++ src/main/java/com/odeyalo/sonata/connect/controller/PlayerController.java
@@ -1,14 +1,16 @@
 package com.odeyalo.sonata.connect.controller;
 
-import com.odeyalo.sonata.connect.dto.*;
-import com.odeyalo.sonata.connect.model.*;
-import com.odeyalo.sonata.connect.service.player.*;
-import com.odeyalo.sonata.connect.service.player.sync.TargetDevices;
+import com.odeyalo.sonata.connect.dto.CurrentlyPlayingPlayerStateDto;
+import com.odeyalo.sonata.connect.dto.PlayResumePlaybackRequest;
+import com.odeyalo.sonata.connect.dto.PlayerStateDto;
+import com.odeyalo.sonata.connect.model.CurrentlyPlayingPlayerState;
+import com.odeyalo.sonata.connect.model.ShuffleMode;
+import com.odeyalo.sonata.connect.model.User;
+import com.odeyalo.sonata.connect.model.Volume;
+import com.odeyalo.sonata.connect.service.player.BasicPlayerOperations;
+import com.odeyalo.sonata.connect.service.player.PlayCommandContext;
 import com.odeyalo.sonata.connect.service.support.mapper.Converter;
-import com.odeyalo.sonata.connect.service.support.mapper.dto.ConnectDeviceRequest2DeviceConverter;
 import com.odeyalo.sonata.connect.service.support.mapper.dto.CurrentPlayerState2PlayerStateDtoConverter;
-import com.odeyalo.sonata.connect.service.support.mapper.dto.Devices2DevicesDtoConverter;
-import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.jetbrains.annotations.NotNull;
 import org.springframework.http.HttpHeaders;
@@ -24,8 +26,6 @@ import reactor.core.scheduler.Schedulers;
 public class PlayerController {
     private final BasicPlayerOperations playerOperations;
     private final CurrentPlayerState2PlayerStateDtoConverter playerState2PlayerStateDtoConverter;
-    private final Devices2DevicesDtoConverter devicesDtoConverter;
-    private final ConnectDeviceRequest2DeviceConverter deviceModelConverter;
     private final Converter<CurrentlyPlayingPlayerState, CurrentlyPlayingPlayerStateDto> currentlyPlayingPlayerStateDtoConverter;
 
     @GetMapping("/state")
@@ -41,13 +41,6 @@ public class PlayerController {
         return playerOperations.currentlyPlayingState(user)
                 .map(state -> ResponseEntity.ok(convertToCurrentlyPlayingStateDto(state)))
                 .defaultIfEmpty(ResponseEntity.noContent().build());
-    }
-
-    @GetMapping("/devices")
-    public Mono<ResponseEntity<?>> getAvailableDevices(User user) {
-        return playerOperations.getDeviceOperations().getConnectedDevices(user)
-                .map(this::convertToAvailableDevicesResponseDto)
-                .map(body -> ResponseEntity.ok().contentType(MediaType.APPLICATION_JSON).body(body));
     }
 
     @PutMapping("/play")
@@ -72,25 +65,6 @@ public class PlayerController {
                 .map(playerState -> default204Response());
     }
 
-    @PutMapping(value = "/device/connect", produces = MediaType.APPLICATION_JSON_VALUE)
-    public Mono<ResponseEntity<?>> addDevice(User user,
-                                             @Valid @RequestBody ConnectDeviceRequest body) {
-        Device device = deviceModelConverter.convertTo(body);
-
-        return playerOperations.getDeviceOperations().addDevice(user, device)
-                .thenReturn(default204Response());
-    }
-
-    @PutMapping(value = "/device/switch", produces = MediaType.APPLICATION_JSON_VALUE)
-    public Mono<ResponseEntity<?>> switchDevices(User user, @RequestBody DeviceSwitchRequest body) {
-        return playerOperations.getDeviceOperations().transferPlayback(
-                        user,
-                        SwitchDeviceCommandArgs.noMatter(),
-                        TargetDeactivationDevices.empty(),
-                        TargetDevices.fromDeviceIds(body.getDeviceIds()))
-                .thenReturn(default204Response());
-    }
-
     @PutMapping(value = "/volume", produces = MediaType.APPLICATION_JSON_VALUE)
     public Mono<ResponseEntity<?>> changePlayerVolume(@NotNull final Volume volume,
                                                       @NotNull final User user) {
@@ -98,22 +72,9 @@ public class PlayerController {
                 .map(it -> default204Response());
     }
 
-    @DeleteMapping(value = "/device")
-    public Mono<ResponseEntity<?>> disconnectDevice(@RequestParam("device_id") String deviceId, User user) {
-        return playerOperations.getDeviceOperations()
-                .disconnectDevice(user, DisconnectDeviceArgs.withDeviceId(deviceId))
-                .thenReturn(default204Response());
-    }
-
     @NotNull
     private CurrentlyPlayingPlayerStateDto convertToCurrentlyPlayingStateDto(CurrentlyPlayingPlayerState state) {
         return currentlyPlayingPlayerStateDtoConverter.convertTo(state);
-    }
-
-    @NotNull
-    private AvailableDevicesResponseDto convertToAvailableDevicesResponseDto(Devices devices) {
-        DevicesDto devicesDto = devicesDtoConverter.convertTo(devices);
-        return AvailableDevicesResponseDto.of(devicesDto);
     }
 
     @NotNull
